update linting rules
