+++ Cargo.toml
@@ -30,21 +30,37 @@ rustdoc.all = "warn"
 branches_sharing_code = "warn"
 clear_with_drain = "warn"
 derive_partial_eq_without_eq = "warn"
+doc_markdown = "warn"
+empty_line_after_doc_comments = "warn"
 empty_line_after_outer_attr = "warn"
+enum_glob_use = "warn"
 equatable_if_let = "warn"
+explicit_into_iter_loop = "warn"
+explicit_iter_loop = "warn"
+flat_map_option = "warn"
 imprecise_flops = "warn"
 iter_on_empty_collections = "warn"
+iter_on_single_items = "warn"
 iter_with_drain = "warn"
+iter_without_into_iter = "warn"
 large_stack_frames = "warn"
+manual_assert = "warn"
 manual_clamp = "warn"
+manual_string_new = "warn"
+match_same_arms = "warn"
 missing_const_for_fn = "warn"
 mutex_integer = "warn"
+naive_bytecount = "warn"
+needless_bitwise_bool = "warn"
+needless_continue = "warn"
 needless_pass_by_ref_mut = "warn"
 nonstandard_macro_braces = "warn"
 or_fun_call = "warn"
 path_buf_push_overwrite = "warn"
 read_zero_byte_vec = "warn"
 redundant_clone = "warn"
+single_char_pattern = "warn"
+string_lit_as_bytes = "warn"
 suboptimal_flops = "warn"
 suspicious_operation_groupings = "warn"
 trailing_empty_array = "warn"
@@ -52,12 +68,14 @@ trait_duplication_in_bounds = "warn"
 transmute_undefined_repr = "warn"
 trivial_regex = "warn"
 tuple_array_conversions = "warn"
+type_repetition_in_bounds = "warn"
 uninhabited_references = "warn"
-uninlined_format_args = "warn"
+unnecessary_struct_initialization = "warn"
 unused_peekable = "warn"
 unused_rounding = "warn"
 use_self = "warn"
 useless_let_if_seq = "warn"
+zero_sized_map_values = "warn"
 
 # These are nursery lints which have findings. Allow them for now. Some are not
 # quite mature enough for use in our codebase and some we don't really want.
@@ -67,20 +85,13 @@ cognitive_complexity = "allow"
 collection_is_never_read = "allow"
 debug_assert_with_mut_call = "allow"
 disallowed_names = "allow"
-empty_line_after_doc_comments = "allow"
 fallible_impl_from = "allow"
 future_not_send = "allow"
-iter_on_single_items = "allow"
 needless_collect = "allow"
 non_send_fields_in_send_ty = "allow"
-octal_escapes = "allow"
-option_if_let_else = "allow"
 redundant_pub_crate = "allow"
 significant_drop_in_scrutinee = "allow"
 significant_drop_tightening = "allow"
-string_lit_as_bytes = "allow"
-type_repetition_in_bounds = "allow"
-unnecessary_struct_initialization = "allow"
 
 [workspace.dependencies]
 alloy = { version = "0.1.2", features = [
